Add TEAMS_EXCLUDING_SELF but keep TEAMS defined similarly
